move constant to a separate file
