+++ src/stories/create-collection/CreateCollection.stories.tsx
@@ -12,7 +12,7 @@ import { MetadataBlockInfoMockRepository } from '../shared-mock-repositories/met
 import { MetadataBlockInfoMockLoadingRepository } from '../shared-mock-repositories/metadata-block-info/MetadataBlockInfoMockLoadingRepository'
 import { MetadataBlockInfoMockErrorRepository } from '../shared-mock-repositories/metadata-block-info/MetadataBlockInfoMockErrorRepository'
 
-import { ROOT_COLLECTION_ALIAS } from '@tests/e2e-integration/shared/collection/CollectionHelper'
+import { ROOT_COLLECTION_ALIAS } from '@tests/e2e-integration/shared/collection/ROOT_COLLECTION_ALIAS'
 
 const meta: Meta<typeof CreateCollection> = {
   title: 'Pages/Create Collection',
