+++ src/stories/shared/add-data-actions/AddDataActionsButton.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { WithI18next } from '../../WithI18next'
 import AddDataActionsButton from '../../../sections/shared/add-data-actions/AddDataActionsButton'
 
-import { ROOT_COLLECTION_ALIAS } from '@tests/e2e-integration/shared/collection/CollectionHelper'
+import { ROOT_COLLECTION_ALIAS } from '@tests/e2e-integration/shared/collection/ROOT_COLLECTION_ALIAS'
 
 const meta: Meta<typeof AddDataActionsButton> = {
   title: 'Sections/Shared/AddDataActions/AddDataActionsButton',
