+++ tests/e2e-integration/shared/collection/CollectionHelper.ts
@@ -50,5 +50,3 @@ export class CollectionHelper extends DataverseApiHelper {
     return collectionResponse
   }
 }
-
-export const ROOT_COLLECTION_ALIAS = 'root'
