+++ tests/e2e-integration/shared/collection/ROOT_COLLECTION_ALIAS.ts
@@ -1,0 +1,1 @@
+export const ROOT_COLLECTION_ALIAS = 'root'
